Move minimal wrapdim functionality to core, remove THTensor include in TensorImpl.
