+++ aten/src/ATen/TensorImpl.cpp
@@ -5,10 +5,9 @@
 #include <ATen/core/optional.h>
 #include <ATen/Context.h>
 #include <ATen/core/Backend.h>
+#include <ATen/core/WrapDimMinimal.h>
 
 #include <ATen/detail/VariableHooksInterface.h>
-
-#include <TH/THTensor.hpp>
 
 namespace at {
